+++ config.py
@@ -8,6 +8,8 @@ def get_config_value(section, key, default=None, is_numeric=False):
                 number = int(DAP.configGetItem(section, key))
                 return number
             except ValueError:
+                from logger import log_it
+                log_it("e", f"Invalid setting for {section}.{key}. Expected an integer, got {DAP.configGetItem(section, key)}")
                 return default
         return DAP.configGetItem(section, key)
     except Exception:
@@ -15,7 +17,7 @@ def get_config_value(section, key, default=None, is_numeric=False):
 
 class Config:
     PLUGIN_NAME = "Cellframe Masternode WebUI"
-    API_TOKEN = str(get_config_value("webui", "api_token", default=False, is_numeric=False))
+    ACCESS_TOKEN = str(get_config_value("webui", "access_token", default=None, is_numeric=False))
     AUTH_BYPASS = get_config_value("webui", "auth_bypass", default=False, is_numeric=False)
     AUTO_UPDATE = get_config_value("webui", "auto_update", default=False, is_numeric=False)
     CACHE_AGE_LIMIT = get_config_value("webui", "cache_age_limit", default=4, is_numeric=True)