+++ templates/scripts/misc.js
@@ -128,6 +128,14 @@ function showChangelogModal() {
     fetch("https://raw.githubusercontent.com/hyttmi/cellframe-masternode-webui/refs/heads/master/CHANGELOG.md")
         .then(response => response.text())
         .then(data => {
+            const emojiMap = {
+                ':sunglasses:': '😎',
+            };
+
+            for (const [shortcode, emoji] of Object.entries(emojiMap)) {
+                data = data.replaceAll(shortcode, emoji);
+            }
+
             document.getElementById("changelog").innerHTML = `
                 <h6>What's New:</h6>
                 <pre>${data}</pre>
@@ -138,6 +146,7 @@ function showChangelogModal() {
             console.error("Failed to load changelog:", error);
         });
 }
+
 
 document.addEventListener("DOMContentLoaded", function () {
     sortCards();